* src/modest-ui-actions.[ch]:         * Implemented ..._on_attachment_clicked, calling view attachment in view.         * Added actions ..._on_attach_file, ..._on_remove_attachments.         * Added support for "select all" action when attachment is focused.         * Added actions ..._view_attachment, ..._save_attachments,           ..._remove_attachments. * src/modest-text-utils.c:         * Now the ..._get_display_size should show correctly a 0 bytes size. * src/maemo/modest-main-window-ui.h:         * Added view, save and remove attachment actions. * src/maemo/modest-msg-edit-window.c:         * Added text_buffer_delete_range handler, to remove image attachments when           they're removed in the body editor.         * ..._add_attachment_clicked handler, to show the add attachment dialog           when user clicks the attachment button. (fixes NB#56314).         * Added ..._attach_file method, to show the dialog to attach one or more           files, and attach them to the message.         * Added ..._remove_attachments method, to remove the selected attachments. * src/maemo/modest-platform.c:         * Now osso_context is a module static pointer, to let platform function use           this.         * Added implementation of modest_platform_activate_file using hildon/osso           mime libraries. * src/maemo/modest-msg-view-window.c:         * ..._update_dimmed now also updates the attachment actions depending on the           current selection.         * Removed some debugging messages.         * Added ..._view_attachment method, to view the currently selected attachment.           If it's a message, it opens a new message view. If it's an attachment, it           calls the default handler.         * Added ..._save_attachments method. This method shows a dialog to select           the folder location to save (or the full path if only one attachment will be           saved).         * Added ..._remove_attachments method (with no implementation currently). * src/maemo/modest-maemo-utils.[ch]:         * Added ..._folder_writable method, that checks if a folder is writable.         * Added ..._file_exists, that checks if a file path corresponds to an existing           file.         * Added ..._create_temp_stream, that creates a temporary file and a corresponding           TnyFsStream. * src/maemo/ui/modest-msg-edit-window-ui.xml:         * Added attach file and remove attachment menu actions. * src/maemo/ui/modest-msg-view-window-ui.xml:         * Added attach file, save attachments and remove attachments menu actions. * src/modest-platform.h:         * New activate file method. * src/gnome/modest-msg-edit-window.c:         * Added new stub functions. * src/gnome/modest-platform.c:         * Added new stub functions. * src/gnome/modest-msg-view-window.c:         * Added new stub functions. * src/widgets/modest-msg-edit-window.h:         * Added methods to attach file and remove attachments. * src/widgets/modest-msg-view.[ch]:         * Remove debugging messages.         * New ..._get_selected_attachments method, that obtains the TnyMimePart's of           current selection in attachments view. * src/widgets/modest-attachment-view.[ch]:         * Removed activated signal. Now activation is completely managed in           ModestAttachmentsView.         * Added support for showing attached messages.         * Removed the button press and button release         * Now labels are not selectable (selection is managed in ModestAttachmentsView).         * Now it inherits from GtkEventBox. * src/widgets/modest-msg-view-window.h:         * Added attachment methods. * src/widgets/modest-attachments-view.[ch]:         * Now it inherits from an event box.         * Minor style fixes.         * Added support for attaching messages.         * New ..._remove_attachment method, to remove attachments corresponding to a           specific TnyMimePart.         * New ..._remove_attachment_by_id, to remove attachments corresponding to a           specific content id.         * Button events handling for rubber banding, selection, and activation.         * Keyboard events handling for selection and activation.         * Internal methods to handle selections: unselect_all, get_att_view_att_coords,           set_selected, select_range.         * Internal methods to handle clipboard selection.         * New ..._get_selection method, that obtains the mime parts of the selected           items.         * New ..._select_all method, that selects all the attachments. * src/widgets/modest-msg-edit-window-ui.h:         * Added actions for attachments. * configure.ac:         * Now maemo target depends on libtinymail-gnomevfs.
pmo-trunk-r1795

+++ src/modest-text-utils.h
@@ -242,7 +242,7 @@ gchar* modest_text_utils_get_display_date (time_t date);
  * Returns: the newly allocated display string for the
  * size in bytes. must be freed.
  */
-gchar * modest_text_utils_get_display_size (guint size);
+gchar * modest_text_utils_get_display_size (guint64 size);
 
 
 /**
